+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/DefaultKubernetesClient.java
@@ -18,7 +18,9 @@ package io.fabric8.kubernetes.client;
 import io.fabric8.kubernetes.client.dsl.FunctionCallable;
 import io.fabric8.kubernetes.client.extended.leaderelection.LeaderElectorBuilder;
 import io.fabric8.kubernetes.client.http.HttpClient;
+import io.fabric8.kubernetes.client.internal.okhttp.OkHttpClientImpl;
 import io.fabric8.kubernetes.client.utils.Serialization;
+import okhttp3.OkHttpClient;
 
 import java.io.InputStream;
 
@@ -38,6 +40,15 @@ public class DefaultKubernetesClient extends BaseKubernetesClient<NamespacedKube
 
   public DefaultKubernetesClient(Config config) {
     super(config);
+  }
+  
+  /**
+   * @deprecated use {@link DefaultKubernetesClient#DefaultKubernetesClient(HttpClient, Config)} instead.
+   * use {@link OkHttpClientImpl#OkHttpClientImpl(OkHttpClient)} to wrap the client.
+   */
+  @Deprecated
+  public DefaultKubernetesClient(OkHttpClient httpClient, Config config) {
+    super(new OkHttpClientImpl(httpClient), config);
   }
   
   public DefaultKubernetesClient(HttpClient httpClient, Config config) {