+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/utils/TokenRefreshInterceptor.java
@@ -17,6 +17,7 @@ package io.fabric8.kubernetes.client.utils;
 
 import io.fabric8.kubernetes.client.Config;
 import io.fabric8.kubernetes.client.http.BasicBuilder;
+import io.fabric8.kubernetes.client.http.HttpClient;
 import io.fabric8.kubernetes.client.http.HttpResponse;
 import io.fabric8.kubernetes.client.http.Interceptor;
 
@@ -30,9 +31,11 @@ public class TokenRefreshInterceptor implements Interceptor {
   public static final String NAME = "TOKEN"; 
   
   private final Config config;
+  private HttpClient.Factory factory;
   
-  public TokenRefreshInterceptor(Config config) {
+  public TokenRefreshInterceptor(Config config, HttpClient.Factory factory) {
     this.config = config;
+    this.factory = factory;
   }
   
   @Override
@@ -47,7 +50,7 @@ public class TokenRefreshInterceptor implements Interceptor {
       }
       Config newestConfig = Config.autoConfigure(currentContextName);
       if (newestConfig.getAuthProvider() != null && newestConfig.getAuthProvider().getName().equalsIgnoreCase("oidc")) {
-        newAccessToken = OpenIDConnectionUtils.resolveOIDCTokenFromAuthConfig(newestConfig.getAuthProvider().getConfig());
+        newAccessToken = OpenIDConnectionUtils.resolveOIDCTokenFromAuthConfig(newestConfig.getAuthProvider().getConfig(), factory.newBuilder());
       } else {
         newAccessToken = newestConfig.getOauthToken();
       }